feature: minor ui fixes

+++ app/src/main/java/kz/ioka/android/iokademoapp/data/OrderRepository.kt
@@ -4,11 +4,12 @@ import kz.ioka.android.iokademoapp.data.local.ProfileDao
 import kz.ioka.android.iokademoapp.data.remote.CheckoutRequestDto
 import kz.ioka.android.iokademoapp.data.remote.CheckoutResponseDto
 import kz.ioka.android.iokademoapp.data.remote.DemoApi
+import java.math.BigDecimal
 import javax.inject.Inject
 
 interface OrderRepository {
 
-    suspend fun checkout(price: Int): CheckoutResponseDto
+    suspend fun checkout(price: BigDecimal): CheckoutResponseDto
 
 }
 
@@ -17,7 +18,7 @@ class OrderRepositoryImpl @Inject constructor(
     private val profileDao: ProfileDao
 ) : OrderRepository {
 
-    override suspend fun checkout(price: Int): CheckoutResponseDto {
+    override suspend fun checkout(price: BigDecimal): CheckoutResponseDto {
         val response = demoApi.checkout(CheckoutRequestDto(price))
 
         response.customerToken?.let {
