+++ app/src/main/java/kz/ioka/android/iokademoapp/data/remote/DemoDto.kt
@@ -2,13 +2,14 @@ package kz.ioka.android.iokademoapp.data.remote
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class ProfileResponseDto(
     @Expose @SerializedName("customer_access_token") val customerToken: String?
 )
 
 data class CheckoutRequestDto(
-    @Expose @SerializedName("price") val price: Int
+    @Expose @SerializedName("price") val price: BigDecimal
 )
 
 data class CheckoutResponseDto(
